Trait hotshot_qc::snarked::circuit::QCKeyAggregateGadget

pub trait QCKeyAggregateGadget<F>where
    F: RescueParameter,{
    // Required methods
    fn check_aggregate_vk<E: EmulationConfig<F>, V: VerKeyVar<E>>(
        &mut self,
        vks: &[V],
        bit_vec: &[BoolVar],
        agg_vk: &V,
        coef: E
    ) -> Result<(), CircuitError>;
    fn check_stake_table_digest<E: EmulationConfig<F>, V: VerKeyVar<E>>(
        &mut self,
        vks: &[V],
        stake_amts: &[Variable],
        digest: Variable
    ) -> Result<(), CircuitError>;
    fn check_threshold(
        &mut self,
        stake_amts: &[Variable],
        bit_vec: &[BoolVar],
        threshold: Variable
    ) -> Result<(), CircuitError>;
}

Plonk circuit gadget for stake key aggregation for quorum certificates.

Required Methods

fn check_aggregate_vk<E: EmulationConfig<F>, V: VerKeyVar<E>>( &mut self, vks: &[V], bit_vec: &[BoolVar], agg_vk: &V, coef: E ) -> Result<(), CircuitError>

Key aggregation circuit

vks - list of stake public keys.

bit_vec - the indicator vector for the quorum set, bit_vec[i] = 1 if i is in the quorum set, o/w bit_vec[i] = 0.

agg_vk - the public aggregated stake key.

coef - the internal curve parameter

fn check_stake_table_digest<E: EmulationConfig<F>, V: VerKeyVar<E>>( &mut self, vks: &[V], stake_amts: &[Variable], digest: Variable ) -> Result<(), CircuitError>

Stake table commitment checking circuit

vk - list of stake public keys.

stake_amts - list of stake amounts for the corresponding stake keys.

digest - the hash of the stake table.

fn check_threshold( &mut self, stake_amts: &[Variable], bit_vec: &[BoolVar], threshold: Variable ) -> Result<(), CircuitError>

Quorum threshold checking circuit

stake_amts - list of stake amounts for the corresponding stake keys.

bit_vec - the indicator vector for the quorum set.

threshold - the public quorum threshold.