Trait hotshot_qc::snarked::circuit::VerKeyVar

pub trait VerKeyVar<E>: Sized + Clone {
    type KeyType: Default;

    // Required methods
    fn native_vars(&self) -> Vec<Variable>;
    fn aggregate_with_selectors<F>(
        circuit: &mut PlonkCircuit<F>,
        keys: &[Self],
        selectors: &[BoolVar],
        coef: E
    ) -> Result<Self, CircuitError>
       where F: PrimeField,
             E: EmulationConfig<F>;
    fn is_equal<F>(
        circuit: &mut PlonkCircuit<F>,
        p0: &Self,
        p1: &Self
    ) -> Result<BoolVar, CircuitError>
       where F: PrimeField,
             E: EmulationConfig<F>;
    fn enforce_equal<F>(
        circuit: &mut PlonkCircuit<F>,
        p0: &Self,
        p1: &Self
    ) -> Result<(), CircuitError>
       where F: PrimeField,
             E: EmulationConfig<F>;
}

Traits for verification keys

Required Associated Types

type KeyType: Default

Required Methods

fn native_vars(&self) -> Vec<Variable>

Returns a list of variables associated with this key variable.

fn aggregate_with_selectors<F>( circuit: &mut PlonkCircuit<F>, keys: &[Self], selectors: &[BoolVar], coef: E ) -> Result<Self, CircuitError>where F: PrimeField, E: EmulationConfig<F>,

Aggregate the verification keys with Boolean selectors.

circuit - associated Plonk circuit.

keys - list of input verification keys.

selectors - list of Boolean selectors.

coef - the internal curve parameter.

Returns an aggregated key variable.

fn is_equal<F>( circuit: &mut PlonkCircuit<F>, p0: &Self, p1: &Self ) -> Result<BoolVar, CircuitError>where F: PrimeField, E: EmulationConfig<F>,

Check whether two input verification key variables are equal.

circuit - associated Plonk circuit.

p0 - first verification key variable.

p1 - second verification key variable.

Returns a Boolean variable indicates whether p0 == p1.

fn enforce_equal<F>( circuit: &mut PlonkCircuit<F>, p0: &Self, p1: &Self ) -> Result<(), CircuitError>where F: PrimeField, E: EmulationConfig<F>,

Enforce that two input verification key variables are equal.

circuit - associated Plonk circuit.

p0 - first verification key variable.

p1 - second verification key variable.